Fix fragile libretime-release downloading
This is just a quick hack to get it working again. Once the rpms
are really stable I don't expect the file to change often and we
can start offering it from github. I plan on automating this
from travis somehow so this hack has to do for the time being.